pyproject.toml
[project]
name = "doc-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "rithvikrthampy", email = "rithvikrthampy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "opencv-python-headless==4.12.0.88",
    "numpy >=2,<2.3.0",
    "scikit-image==0.25.2",
    "Pillow==11.3.0",
    "rembg (>=2.0.67,<3.0.0)",
    "onnxruntime-gpu (>=1.22.0,<2.0.0)",
    "onnxruntime (>=1.22.1,<2.0.0)",
    "deskew (>=1.5.3,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "fastapi (>=0.104.0,<1.0.0)",
    "uvicorn[standard] (>=0.24.0,<1.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)",
    "onnxruntime-directml (>=1.22.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "doc_analysis", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
